refactor: remove vesting related files/functions from stream, vesting init
