+++ the-graph/subgraphs/furo-stream/src/functions/user.ts
@@ -1,5 +1,5 @@
 import { Address, ethereum } from '@graphprotocol/graph-ts'
-import { User } from '../schema'
+import { User } from '../../generated/schema'
 import { increaseUserCount } from './furo'
