+++ the-graph/subgraphs/furo-vesting/README.md
@@ -1,0 +1,11 @@
+# Furo Vesting Subgraph
+
+## Unit testing
+
+### Run
+
+```sh
+yarn graph test <ONE-OR-MORE-TEST-NAMES>
+
+yarn graph test Vesting
+```
