+++ the-graph/subgraphs/furo-vesting/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@sushiswap/furo-vesting-subgraph",
+  "author": "Ola Stenberg <hello@olastenberg.com>",
+  "license": "MIT",
+  "repository": {
+    "url": "sushiswap/subgraphs",
+    "directory": "subgraphs/furo-vesting"
+  },
+  "files": [
+    "generated"
+  ],
+  "scripts": {
+    "codegen": "graph codegen",
+    "build": "graph build",
+    "deploy": "graph deploy --node https://api.thegraph.com/deploy/ sushiswap/furo-vesting",
+    "create-local": "graph create --node http://localhost:8020/ sushiswap/furo-vesting",
+    "remove-local": "graph remove --node http://localhost:8020/ sushiswap/furo-vesting",
+    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 sushiswap/furo-vesting"
+  },
+  "devDependencies": {
+    "@graphprotocol/graph-cli": "^0.27.0",
+    "@graphprotocol/graph-ts": "^0.24.1",
+    "matchstick-as": "^0.4.2"
+  }
+}
