+++ the-graph/subgraphs/furo-vesting/src/constants/index.ts
@@ -1,0 +1,9 @@
+export const ONGOING = 'ONGOING'
+export const CANCELLED = 'CANCELLED'
+export const EXPIRED = 'EXPIRED'
+
+export const DEPOSIT = 'DEPOSIT'
+export const WITHDRAWAL = 'WITHDRAWAL'
+export const DISBURSEMENT = 'DISBURSEMENT'
+
+export const FURO = 'Furo'
