+++ the-graph/subgraphs/furo-vesting/src/functions/user.ts
@@ -1,0 +1,19 @@
+// import { Address, ethereum } from '@graphprotocol/graph-ts'
+// import { User } from '../schema'
+// import { increaseUserCount } from './furo'
+
+
+// export function getOrCreateUser(id: Address, event: ethereum.Event): User {
+//   let user = User.load(id.toHex())
+
+//   if (user === null) {
+//     user = new User(id.toHex())
+//     user.createdAtBlock = event.block.number
+//     user.createdAtTimestamp = event.block.timestamp
+//     increaseUserCount()
+//   }
+
+//   user.save()
+
+//   return user as User
+// }
